shift main code to main.py instead of parser, some refinments
