+++ setup.py
@@ -33,7 +33,7 @@ if __name__ == "__main__":
                      'Programming Language :: Python :: 3'],
         entry_points='''
             [console_scripts]
-            deltabot=deltabot.parser:main
+            deltabot=deltabot.main:main
             [pytest11]
             deltabot.pytestplugin=deltabot.pytestplugin
         ''',
