+++ src/deltabot/__main__.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 
-from .parser import main
+from .main import main
 
 
 if __name__ == "__main__":
